reorganize testing
